+++ core/bytecode/src/compiler.rs
@@ -1207,7 +1207,7 @@ impl Compiler {
                                     &[*target_register, iter_register, i as u8],
                                 );
                             } else {
-                                self.push_op(IterNext, &[*target_register, iter_register, 0, 0]);
+                                self.push_op(IterUnpack, &[*target_register, iter_register]);
                             }
                             // The register was reserved before the RHS was compiled, and now it
                             // needs to be committed.
@@ -1223,7 +1223,7 @@ impl Compiler {
                             if rhs_is_temp_tuple {
                                 self.push_op(TempIndex, &[value_register, iter_register, i as u8]);
                             } else {
-                                self.push_op(IterNext, &[value_register, iter_register, 0, 0]);
+                                self.push_op(IterUnpack, &[value_register, iter_register]);
                             }
 
                             self.compile_value_export(*id_index, value_register)?;
@@ -1248,7 +1248,7 @@ impl Compiler {
                     if rhs_is_temp_tuple {
                         self.push_op(TempIndex, &[value_register, iter_register, i as u8]);
                     } else {
-                        self.push_op(IterNext, &[value_register, iter_register, 0, 0]);
+                        self.push_op(IterUnpack, &[value_register, iter_register]);
                     }
 
                     self.compile_lookup(
@@ -1273,7 +1273,7 @@ impl Compiler {
                         if rhs_is_temp_tuple {
                             self.push_op(TempIndex, &[value_register, iter_register, i as u8]);
                         } else {
-                            self.push_op(IterNext, &[value_register, iter_register, 0, 0]);
+                            self.push_op(IterUnpack, &[value_register, iter_register]);
                         }
 
                         self.push_op(SequencePush, &[result.register, value_register]);
@@ -3734,23 +3734,24 @@ impl Compiler {
                 // e.g. for a, b, c in list_of_lists()
                 // e.g. for key, value in map
 
-                // A temporary register for the iterator output.
-                // Args are unpacked from the temp register
+                // A temporary register for the iterator's output.
+                // Args are unpacked via iteration from the temp register
                 let temp_register = self.push_register()?;
 
                 self.push_op_without_span(IterNextTemp, &[temp_register, iterator_register]);
                 self.push_loop_jump_placeholder()?;
 
-                for (i, arg) in args.iter().enumerate() {
+                self.push_op_without_span(MakeIterator, &[temp_register, temp_register]);
+
+                for arg in args.iter() {
                     match &ast.node(*arg).node {
                         Node::Id(id) => {
                             let arg_register = self.assign_local_register(*id)?;
-                            self.push_op_without_span(
-                                TempIndex,
-                                &[arg_register, temp_register, i as u8],
-                            );
+                            self.push_op_without_span(IterUnpack, &[arg_register, temp_register]);
                         }
-                        Node::Wildcard(_) => {}
+                        Node::Wildcard(_) => {
+                            self.push_op_without_span(IterNextQuiet, &[temp_register, 0, 0]);
+                        }
                         unexpected => {
                             return compiler_error!(
                                 self,