+++ core/runtime/src/value_tuple.rs
@@ -34,6 +34,32 @@ impl ValueTuple {
     pub fn is_hashable(&self) -> bool {
         self.iter().all(Value::is_hashable)
     }
+
+    /// Removes and returns the first value in the tuple
+    ///
+    /// The internal bounds of the tuple are adjusted to 'remove' the first element;
+    /// no change is made to the underlying tuple data.
+    pub fn pop_front(&mut self) -> Option<Value> {
+        if let Some(value) = self.first().cloned() {
+            self.bounds.start += 1;
+            Some(value)
+        } else {
+            None
+        }
+    }
+
+    /// Removes and returns the last value in the tuple
+    ///
+    /// The internal bounds of the tuple are adjusted to 'remove' the first element;
+    /// no change is made to the underlying tuple data.
+    pub fn pop_back(&mut self) -> Option<Value> {
+        if let Some(value) = self.last().cloned() {
+            self.bounds.end -= 1;
+            Some(value)
+        } else {
+            None
+        }
+    }
 }
 
 impl Deref for ValueTuple {